requirements.txt
APScheduler==3.8.1
flask-apscheduler==1.12.2
influxdb==5.3.0
Flask==2.0.2
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-crontab==0.1.2
gevent==20.5.0
paramiko==2.6.0
prettytable==2.4.0
psycopg2==2.7.2
# for windows
# psycopg2-binary==2.9.2
python-dateutil==2.8.0
python-jenkins==1.7.0
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.26.0
SQLAlchemy==1.4.26
python-dotenv==0.19.2
pydantic==1.8.2
# format
isort==5.10.1
flake8==4.0.1
black==21.12b0